+++ src/components/About/About.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './About.module.css';
+
+
+const About = () => (
+  <h2 className={styles.title}>Elvira Snezhkova</h2>
+);
+
+
+export default About;
