+++ src/components/App/App.js
@@ -1,79 +1,35 @@
 import React from 'react';
-import InputItem from '../InputItem/InputItem';
-import ItemList from '../ItemList/ItemList';
-import Footer from '../Footer/Footer';
+import MenuList from '@material-ui/core/MenuList';
+import MenuItem from '@material-ui/core/MenuItem';
+import Paper from '@material-ui/core/Paper';
+import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
+
+import Todo from '../Todo/Todo';
+import About from '../About/About';
+import Contacts from '../Contacts/Contacts';
 import styles from './App.module.css';
 
-class App extends React.Component {
-  state = {
-    todoItems: [
-      {
-        value: 'Сделать зарядку',
-        isDone: true,
-        id: 1
-      },
-      {
-        value: 'Забрать пальто из химчистки',
-        isDone: false,
-        id: 2
-      },
-      {
-        value: 'Отвезти кота в вет клинику',
-        isDone: false,
-        id: 3
-      }
-    ],
 
-    count: 3
-  };
 
-  onClickDone = id => {
-    const newItemList = this.state.todoItems.map(item => {
-      let newItem = { ...item };
+const App =  () => (
+     <Router>
+        <div>
+          <Paper className={styles.menu}>
+          <MenuList id="simple-menu" autoWidth={false}>
+            <Link to ='/'><MenuItem className={styles.item} >Обо Мне</MenuItem></Link> 
+            <Link to ='/contacts'><MenuItem className={styles.item} >Контакты</MenuItem></Link> 
+            <Link to ='/todo'><MenuItem className={styles.item} >Мои дела</MenuItem></Link>
+          </MenuList>
+          </Paper>
+
+          <div className={styles.wrapper}>
+            <Route path='/' exact component={About} />
+            <Route path='/contacts' exact component={Contacts} />
+            <Route path='/todo' exact component={Todo} />
+          </div>
+        </div>
+     </Router> 
   
-      if (item.id === id) {
-        newItem.isDone = !item.isDone;
-      }
-
-      return newItem;
-    });
-
-    this.setState({ todoItems: newItemList });
-  };
-
-  onClickDelete = id => {
-    const newItemList = this.state.todoItems.filter(item => item.id !== id);
-    this.setState({ todoItems: newItemList });
-    this.setState({count: newItemList.length});
-  }
-
-  onClickAdd = value => {
-      const newItemlist = [
-        ...this.state.todoItems,
-        {
-          value,
-          isDone: false,
-          id: this.state.count + 1 
-        }
-      ];
-      const newCount = newItemlist.length; 
-
-      this.setState({ todoItems: newItemlist,  count: newCount});
-  };
-
-  render() {    
-    return (
-    <div className={styles.wrapper}>
-      <h2 className={styles.title}>Todos</h2>
-      <InputItem onClickAdd={this.onClickAdd} />
-      <ItemList 
-        todoItems={this.state.todoItems} 
-        onClickDone={this.onClickDone} 
-        onClickDelete={this.onClickDelete} />
-      <Footer count={this.state.count} />
-    </div>
-    );
-  }
-};
+);
 
 export default App;
