+++ src/components/App/App.module.css
@@ -14,16 +14,23 @@ body {
 
 .wrapper{
   background: #fff;
-  max-width: 850px;
+  max-width: 1000px;
   width: 100%;
-  margin: 120px auto;
+  margin: 30px auto;
   padding: 25px;
   border-radius: 5px;
   box-shadow: 0px 10px 15px rgba(0,0,0,0.1);
 }
 
-.title {
-  font-size: 30px;
-  font-weight: 600;
-  margin-bottom: 20px;
+.menu {
+  margin: 0 32px 16px 0;
+  max-width: 1000px;
+  height: 50px;
+  margin: 30px auto;
+}
+
+.item {
+  display: inline;
+  float: left;
+  width: 7%;
 }
