+++ src/components/Contacts/Contacts.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import styles from './Contacts.module.css';
+
+
+const Contact = () => (
+  <h1  className={styles.title}>Contact</h1>
+);
+
+
+export default Contact;
