+++ src/components/Contacts/Contacts.module.css
@@ -1,0 +1,5 @@
+.title {
+  font-size: 30px;
+  font-weight: 600;
+  margin-bottom: 20px;
+}
